[REF] point_of_sale: move payment_screen related method to it
Move a method that was in pos_store directly to payment_screen since
its only related to it
